security_inspector/requirements.txt
PyQt6
psutil
requests
pywin32
fpdf2
python-dotenv
